ComplaintSystemMainCourseApp/requirements.txt
alembic==1.9.4
alembic-autogen-check==1.1.1
anyio==3.6.2
asyncclick==8.1.3.4
asyncpg==0.27.0
attrs==22.2.0
black==23.1.0
certifi==2022.12.7
cffi==1.15.1
click==8.1.3
cryptography==39.0.1
databases==0.7.0
dnspython==2.3.0
ecdsa==0.18.0
email-validator==1.3.1
exceptiongroup==1.1.0
fastapi==0.92.0
flake8==6.0.0
greenlet==2.0.2
h11==0.14.0
httpcore==0.16.3
httptools==0.5.0
httpx==0.23.3
idna==3.4
iniconfig==2.0.0
isort==5.12.0
itsdangerous==2.1.2
Jinja2==3.1.2
Mako==1.2.4
MarkupSafe==2.1.2
mccabe==0.7.0
mypy==1.0.1
mypy-extensions==1.0.0
orjson==3.8.6
packaging==23.0
passlib==1.7.4
pathspec==0.11.0
platformdirs==3.0.0
pluggy==1.0.0
psycopg2==2.9.5
pyasn1==0.4.8
pycodestyle==2.10.0
pycparser==2.21
pydantic==1.10.5
pyflakes==3.0.1
pytest==7.2.1
python-decouple==3.7
python-dotenv==0.21.1
python-jose==3.3.0
python-multipart==0.0.5
PyYAML==6.0
rfc3986==1.5.0
rsa==4.9
six==1.16.0
sniffio==1.3.0
SQLAlchemy==1.4.46
starlette==0.25.0
tomli==2.0.1
typing_extensions==4.5.0
ujson==5.7.0
uvicorn==0.20.0
uvloop==0.17.0
watchfiles==0.18.1
websockets==10.4
